Fix int32 overflow in PeptideProfile.randint on Windows
PeptideProfile uses max_intensity=1e10, which exceeds int32's maximum. On
Windows NumPy's default integer is int32, so RandomState.randint raised
"high is out of bounds for int32". Pin dtype=np.int64, which matches the
default already used on Linux/macOS, so values are unchanged there.

+++ directlfq/test_utils.py
@@ -51,7 +51,7 @@ class PeptideProfile():
 
     def _get_single_peptide_profile_template(self):
         rs = RandomState(MT19937(SeedSequence(42312)))
-        return rs.randint(low=self._min_intensity, high=self._max_intensity,size=self._num_samples)
+        return rs.randint(low=self._min_intensity, high=self._max_intensity, size=self._num_samples, dtype=np.int64)
 
     def _scale_profile_vector(self):
         self.peptide_profile_vector = self.peptide_profile_vector*self._systematic_peptide_shift
